fix(generate): split catalog_snapshot O_NOFOLLOW for Windows compat
Windows lacks syscall.O_NOFOLLOW. Cross-compile to GOOS=windows
fails since Plan 32 (#80). Split openSnapshotFile into platform
files: unix variant keeps O_NOFOLLOW symlink defense; windows
variant uses plain OpenFile.

Unblocks v0.4.0 GoReleaser cross-compile (release run 25313754104
failed at undefined: syscall.O_NOFOLLOW).
